[refactor] frameExtractor: ui 레이어 이동 + seek 타임아웃·URL 정리
- frameExtractor.ts를 model/ → ui/로 이동. video/canvas DOM I/O에 의존해
  model(순수 로직, 테스트 대상) 규칙을 위반하므로 유일한 사용처
  useKeyFrameAnalysis와 같은 ui 레이어로 이동
  - jest.config의 frameExtractor 커버리지 예외 제거(더 이상 model 아님)
- seek 대기에 타임아웃·error 처리 추가(waitForSeek): seeked가 끝내 오지
  않을 때 영구 대기하던 위험 제거
- objectURL 해제를 try/finally로 감싸 예외 경로에서도 누수 방지
